+++ src/app/company/[id]/manage_reservation/_styles/TotalReservationList.module.css
@@ -1,0 +1,48 @@
+.input-wrapper {
+  position: relative;
+  display: flex;
+  /* align-items: center; */
+}
+
+.input {
+  width: 300px;
+  border: none;
+  border-radius: 74px;
+  padding: 0.5rem;
+  color: #adb0b7;
+  background-color: #e4f6df;
+}
+
+.total-menu {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 1.5rem;
+}
+
+.search-btn {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: absolute;
+  top: 3px;
+  bottom: 3px;
+  right: 8px;
+  width: 1.5rem;
+  height: 1.5rem;
+  border-radius: 50%;
+  background-color: #ffffff;
+}
+
+.table {
+  width: 900px;
+  border-collapse: collapse;
+  border-spacing: 0;
+}
+
+.th {
+  font-weight: 600;
+  text-align: left;
+  border-bottom: 1px solid #dddddd;
+  padding: 0.5rem;
+}
